Factored out diff functionality from the CatalogMergeTool class
CatalogMergeTool now extends CatalogDiffTool, which can later be used
in the swissknife_diff command.
