+++ cvmfs/receiver/commit_processor.cc
@@ -56,10 +56,11 @@ CommitProcessor::Result CommitProcessor::Process(
   }
 
   CatalogMergeTool merge_tool(repo_name, old_root_hash_str, new_root_hash_str,
-                              manifest->catalog_hash().ToString(true));
+                              manifest->catalog_hash().ToString(true),
+                              "/tmp/cvmfs_receiver_merge");
 
   shash::Any resulting_root_hash;
-  if (!merge_tool.Merge(&resulting_root_hash)) {
+  if (!merge_tool.Run(&resulting_root_hash)) {
     LogCvmfs(kLogCvmfs, kLogStderr, "Catalog merge failed");
     return kMergeError;
   }
